processor/requirements.txt
weaviate-client==4.11.1
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.8.1